% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slippymath.R
\name{compose_tile_grid}
\alias{compose_tile_grid}
\title{compose_tile_grid}
\usage{
compose_tile_grid(tile_grid, images)
}
\arguments{
\item{tile_grid}{a tile_grid object, likely returned from `bbox_to_tile_grid`}

\item{images}{a list of character strings defining paths to images. Matched to tiles in tile_grid based on list position.}
}
\value{
a spatially referenced raster.
}
\description{
Compose a list of images using tile_grid data.
}
\details{
Given a tile_grid object and a list of images, compose the images into a
single spatially referenced RasterBrick object.

The list of images is assumed to be in corresponding order to the tiles in
the tile_grid object.

The returned object uses the Web Mercator projection, EPSG:3857, which is
the native crs of the tiles.
}
