% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slippymath.R
\name{latlon_to_tilenum}
\alias{latlon_to_tilenum}
\title{latlon_to_tilenum}
\usage{
latlon_to_tilenum(lat_deg, lon_deg, zoom)
}
\arguments{
\item{lat_deg}{degrees latitude for point}

\item{lon_deg}{degrees longitude for point}

\item{zoom}{zoom level for tile calculation. Increasing zoom increases the
number of tiles.}
}
\value{
a list containing `x` and `y` - the tile numbers.
}
\description{
Convert latitude and longitude to slippy tile numbers
}
\details{
Returns the Open Street Map slippy map tile numbers (x, y) the
supplied latitude and longitude fall on, for a given zoom level.

The point specified by `lat_deg` and `lon_deg` is assumed to be in ESPG:4326
coordinate reference system.
}
