% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercator.R
\name{lonlat_to_merc}
\alias{lonlat_to_merc}
\alias{merc_to_lonlat}
\title{Transform between spherical Mercator and longitude/latitude}
\usage{
lonlat_to_merc(ll)

merc_to_lonlat(xy)
}
\arguments{
\item{ll}{matrix of longitude / latitude}

\item{xy}{matrix of x / y Mercator}
}
\value{
matrix of coordinates transformed forward or inverse
}
\description{
Transform between spherical Mercator and longitude/latitude
}
