% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercator.R
\name{merc_truncate}
\alias{merc_truncate}
\title{merc_truncate}
\usage{
merc_truncate(xy)
}
\arguments{
\item{xy}{a matrix of Mercator XY points.}
}
\value{
a matrix of XY points.
}
\description{
Truncate coordinate to Mercator extent.
}
\details{
If a point in m lies outside the Mercator extent, this function can be used
to truncate it to the boundary of the extent.
}
\examples{
stray <- matrix(c(20037509,
                   -2918162.223463),
                 nrow = 1)

merc_truncate(stray)

}
