% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{raster_to_png}
\alias{raster_to_png}
\title{raster_to_png}
\usage{
raster_to_png(tile_raster, file_path)
}
\arguments{
\item{tile_raster}{the raster to write to PNG}

\item{file_path}{the path to write the raster}
}
\value{
nothing.
}
\description{
Write a raster to PNG
}
\details{
This function is a convenience wrapper for
writing rasters in PNG format.
}
