% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slippymath.R
\name{tg_composite}
\alias{tg_composite}
\title{tg_composite}
\usage{
tg_composite(tile_grid, images)
}
\arguments{
\item{tile_grid}{a tile_grid object, likely returned from `bb_to_tg`}

\item{images}{a list of character strings defining paths to images. Matched to tiles in tile_grid based on list position.}
}
\value{
a spatially referenced raster.
}
\description{
Composite a list of images using tile_grid data.
}
\details{
Given a tile_grid object and a list of images, composite the images into a spatially referenced RasterBrick object.

The list of images is assumed to be in corresponding order to the tiles in
the tile_grid object.

The returned object uses the Web Mercator projection, EPSG:3857, which is
the native crs of the tiles.
}
