% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slippymath.R
\name{tile_bb}
\alias{tile_bb}
\title{tile_bb}
\usage{
tile_bb(x, y, zoom)
}
\arguments{
\item{x}{slippy map tile x number}

\item{y}{slippy map tile y number}

\item{zoom}{zoom level for tile}
}
\value{
an sf bbox object.
}
\description{
Calculate the bounding box for a tile in latitude and longitude
}
\details{
Given a slippy maps tile specified by `x`, `y`, and `zoom`, return the
an `sf` bounding box object for the tile in degrees latitude and longitude using the EPSG:4326
coordinate reference system.
}
