% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slippymath.R
\name{tile_bbox}
\alias{tile_bbox}
\title{tile_bbox}
\usage{
tile_bbox(x, y, zoom)
}
\arguments{
\item{x}{slippy map tile x number}

\item{y}{slippy map tile y number}

\item{zoom}{zoom level for tile}
}
\value{
an sf bbox object.
}
\description{
Calculate the bounding box for a tile in latitude and longitude
}
\details{
Given a slippy maps tile specified by `x`, `y`, and `zoom`, return the
an `sf` bounding box object for the tile with units in metres using the
EPSG:3857 coordinate reference system (Web Mercator).
}
\examples{
## return an sf style bbox object in with epsg and proj4string
tile_bbox(x = 30304, y = 18929, zoom = 15)
}
