% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slippymath.R
\name{tilenum_to_lonlat}
\alias{tilenum_to_lonlat}
\title{tilenum_to_lonlat}
\usage{
tilenum_to_lonlat(x, y, zoom)
}
\arguments{
\item{x}{slippy map tile number in x domain (left to right)}

\item{y}{slippy map tile number in y domain (top to bottom)}

\item{zoom}{the zoom level for the calculation. Increasing zoom increases
the number of tiles.}
}
\value{
a list containing `lat` and `lon` - latitude and longitude.
}
\description{
Convert slippy map tiles numbers to latitude and longitude
}
\details{
Returns the latitude and longitude of the top left corner of a slippy map tile
specified by `x`, `y` for a given zoom level.
}
\examples{
tilenum_to_lonlat(
 x = 70406,
 y = 42987,
 zoom = 17
)
}
