% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercator.R
\name{within_merc_extent}
\alias{within_merc_extent}
\title{within_mercator_extent}
\usage{
within_merc_extent(xy)
}
\arguments{
\item{xy}{a matrix of Mercator xy coordinates.}
}
\value{
TRUE or FALSE
}
\description{
Are points in meters within Mercator extent?
}
\details{
When doing maths with Mercator coordinates in m, you can end up outside the
Mercator extent with an undefined coordinate. This function returns true if
all xy lie within the Mercator extent.
}
\examples{
stray <- matrix(c(20037509,
                   -2918162.223463),
                 nrow = 1)

within_merc_extent(stray)
}
